requirements.txt
python-telegram-bot==21.6
aiosqlite==0.20.0
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
